clinical_feedback_app/requirements.txt

streamlit
openai>=1.0.0
psycopg2-binary
python-dotenv
docx2txt
PyPDF2
fpdf
bcrypt
